requirements.txt
PyQt6>=6.5,<7
PyQt6-WebEngine>=6.5,<7
rdflib>=6.3
owlready2>=0.45
pyvis>=0.3
networkx>=3.0
spacy>=3.5
spacy-lookups-data
google-generativeai>=0.3
python-dotenv>=1.0
